Moved validate up to the common section
