+++ x-pack/plugins/lists/server/routes/find_list_item_route.ts
@@ -7,12 +7,8 @@
 import { IRouter } from 'kibana/server';
 
 import { LIST_ITEM_URL } from '../../common/constants';
-import {
-  buildRouteValidation,
-  buildSiemResponse,
-  transformError,
-  validate,
-} from '../siem_server_deps';
+import { buildRouteValidation, buildSiemResponse, transformError } from '../siem_server_deps';
+import { validate } from '../../common/siem_common_deps';
 import { findListItemSchema, foundListItemSchema } from '../../common/schemas';
 import { decodeCursor } from '../services/utils';
