+++ x-pack/plugins/lists/server/routes/import_list_item_route.ts
@@ -9,12 +9,8 @@ import { Readable } from 'stream';
 import { IRouter } from 'kibana/server';
 
 import { LIST_ITEM_URL } from '../../common/constants';
-import {
-  buildRouteValidation,
-  buildSiemResponse,
-  transformError,
-  validate,
-} from '../siem_server_deps';
+import { buildRouteValidation, buildSiemResponse, transformError } from '../siem_server_deps';
+import { validate } from '../../common/siem_common_deps';
 import { importListItemQuerySchema, importListItemSchema, listSchema } from '../../common/schemas';
 
 import { getListClient } from '.';
