+++ x-pack/plugins/lists/server/routes/read_list_item_route.ts
@@ -7,13 +7,9 @@
 import { IRouter } from 'kibana/server';
 
 import { LIST_ITEM_URL } from '../../common/constants';
-import {
-  buildRouteValidation,
-  buildSiemResponse,
-  transformError,
-  validate,
-} from '../siem_server_deps';
+import { buildRouteValidation, buildSiemResponse, transformError } from '../siem_server_deps';
 import { listItemArraySchema, listItemSchema, readListItemSchema } from '../../common/schemas';
+import { validate } from '../../common/siem_common_deps';
 
 import { getListClient } from '.';
