+++ x-pack/plugins/lists/server/siem_server_deps.ts
@@ -17,5 +17,4 @@ export {
   createBootstrapIndex,
   getIndexExists,
   buildRouteValidation,
-  validate,
 } from '../../security_solution/server';
