+++ x-pack/plugins/security_solution/server/lib/detection_engine/routes/rules/validate.test.ts
@@ -4,10 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import * as t from 'io-ts';
-
 import {
-  validate,
   transformValidate,
   transformValidateFindAlerts,
   transformValidateBulkError,
@@ -119,26 +116,6 @@ describe('validate', () => {
 
   afterAll(() => {
     unSetFeatureFlagsForTestsOnly();
-  });
-
-  describe('validate', () => {
-    test('it should do a validation correctly', () => {
-      const schema = t.exact(t.type({ a: t.number }));
-      const payload = { a: 1 };
-      const [validated, errors] = validate(payload, schema);
-
-      expect(validated).toEqual(payload);
-      expect(errors).toEqual(null);
-    });
-
-    test('it should do an in-validation correctly', () => {
-      const schema = t.exact(t.type({ a: t.number }));
-      const payload = { a: 'some other value' };
-      const [validated, errors] = validate(payload, schema);
-
-      expect(validated).toEqual(null);
-      expect(errors).toEqual('Invalid value "some other value" supplied to "a"');
-    });
   });
 
   describe('transformValidate', () => {
